+++ app/(gcforms)/[locale]/(form administration)/forms/components/Invitations/index.tsx
