+++ app/(gcforms)/[locale]/(form administration)/forms/page.tsx
@@ -10,7 +10,7 @@ import { ResumeEditingForm } from "./components/ResumeEditingForm";
 import { getAllTemplatesForUser, TemplateOptions } from "@lib/templates";
 import { DeliveryOption } from "@lib/types";
 import { getOverdueTemplateIds } from "@lib/overdue";
-import { Invitations } from "./components/server/Invitations";
+import { Invitations } from "./components/Invitations";
 
 export type FormsTemplate = {
   id: string;
@@ -81,7 +81,7 @@ export default async function Page({
     );
 
     return (
-      <div className="center mx-auto w-[980px] bg-gray-soft">
+      <div className="mx-auto w-[980px]">
         <h1 className="mb-8 border-b-0">{t("title")}</h1>
         <Invitations />
         <div className="flex w-full justify-between">
